refactor: パッケージ構造を最適化
